# docs/openapi.yaml
openapi: "3.1.0"
info:
  title: VN Fashion Shop API
  description: |
    RESTful API cho nen tang thuong mai dien tu thoi trang Viet Nam.
    RESTful API for the Vietnamese fashion e-commerce platform.

    ## Xac thuc / Authentication
    Su dung Bearer JWT token trong header `Authorization`.
    Use Bearer JWT tokens in the `Authorization` header.

    ## Phan trang / Pagination
    Cac endpoint tra ve danh sach su dung query params `page` va `page_size`.
    List endpoints use `page` and `page_size` query parameters.

    ## Ma loi / Error Codes
    - `400` -- Bad Request
    - `401` -- Unauthorized (chua xac thuc)
    - `403` -- Forbidden (khong co quyen)
    - `404` -- Not Found
    - `422` -- Validation Error
    - `500` -- Internal Server Error
  version: "1.0.0"
  contact:
    name: VN Fashion Shop Team
    email: dev@vnfashion.vn
servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://api.vnfashion.vn
    description: Production
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token nhan duoc tu endpoint /auth/login
  schemas:
...
    ErrorResponse:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
          example: "Khong tim thay san pham"
    ValidationError:
      type: object
      required: [detail]
      properties:
        detail:
          type: array
          items:
            type: object
            properties:
              loc:
                type: array
                items:
                  oneOf:
                    - type: string
                    - type: integer
              msg:
                type: string
              type:
                type: string

    # ---- Auth ----
...
    CartResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItemResponse"
        total_items:
          type: integer
          example: 3
        subtotal:
          type: number
          format: double
          example: 3600000
        currency:
          type: string
          example: "VND"
    CartItemResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        variant:
          $ref: "#/components/schemas/CartItemVariant"
        quantity:
          type: integer
          example: 2
        line_total:
          type: number
          format: double
          example: 2400000
    CartItemVariant:
      type: object
      properties:
        id:
          type: string
          format: uuid
        product_name:
          type: string
          example: "Ao Dai Hoa Sen"
        sku:
          type: string
          example: "ADHS-DO-M"
        size:
          type: string
          example: "M"
        color:
          type: string
          example: "Do"
        price:
          type: number
          format: double
          example: 1200000
        image_url:
          type: string
        stock_quantity:
          type: integer
          example: 25
    AddToCartRequest:
      type: object
      required: [variant_id, quantity]
      properties:
        variant_id:
          type: string
          format: uuid
        quantity:
          type: integer
          minimum: 1
          example: 1
    UpdateCartItemRequest:
      type: object
      required: [quantity]
      properties:
        quantity:
          type: integer
          minimum: 0
          example: 2

    # ---- Checkout ----
    CheckoutRequest:
      type: object
      required: [shipping_address_id, payment_method]
      properties:
        shipping_address_id:
          type: string
          format: uuid
        payment_method:
          type: string
          enum: [stripe, cod]
          example: "stripe"
        note:
          type: string
          example: "Giao hang truoc 5 gio chieu"
          nullable: true
        coupon_code:
          type: string
          nullable: true
    CheckoutResponse:
      type: object
      properties:
        order_id:
          type: string
          format: uuid
        order_number:
          type: string
          example: "VNF-20260101-0001"
        total_amount:
          type: number
          format: double
          example: 1230000
        payment_method:
          type: string
          example: "stripe"
        stripe_client_secret:
          type: string
          nullable: true
          description: "Stripe PaymentIntent client_secret (chi tra ve khi payment_method=stripe)"

    # ---- Orders ----
...
paths:
...
  /cart:
    get:
      tags: [Cart]
      summary: Xem gio hang / View cart
      operationId: getCart
      responses:
        "200":
          description: Thong tin gio hang
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CartResponse"
        "401":
          description: Chua xac thuc
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /cart/items:
    post:
      tags: [Cart]
      summary: Them san pham vao gio hang / Add item to cart
      operationId: addToCart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddToCartRequest"
      responses:
        "200":
          description: Them thanh cong
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CartResponse"
        "400":
          description: San pham het hang hoac so luong vuot qua ton kho
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                detail: "San pham het hang"
        "401":
          description: Chua xac thuc
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Khong tim thay bien the san pham
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /cart/items/{item_id}:
    patch:
      tags: [Cart]
      summary: Cap nhat so luong / Update quantity
      operationId: updateCartItem
      parameters:
        - name: item_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCartItemRequest"
      responses:
        "200":
          description: Cap nhat thanh cong
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CartResponse"
        "400":
          description: So luong vuot qua ton kho
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Chua xac thuc
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Khong tim thay mat hang trong gio
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      tags: [Cart]
      summary: Xoa san pham khoi gio / Remove item from cart
      operationId: removeCartItem
      parameters:
        - name: item_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Xoa thanh cong
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CartResponse"
        "401":
          description: Chua xac thuc
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Khong tim thay mat hang trong gio
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /checkout:
    post:
      tags: [Checkout]
      summary: Tao don hang va thanh toan / Create order and pay
      operationId: checkout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutRequest"
      responses:
        "201":
          description: Tao don hang thanh cong
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutResponse"
        "400":
          description: Gio hang trong hoac san pham het hang
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                detail: "Gio hang dang trong"
        "401":
          description: Chua xac thuc
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Du lieu khong hop le
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"

  # ======== Orders ========
...